+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+**3.0.3** (April 27, 2015)
+
 * Fix static asset mtime fallback
 * Only warn when specified asset version can not be loaded.